Rename file to match #include directive
Mismatch of case between filename and #include directive causes compilation to fail in case sensitive operating systems such as Linux. Rather than correct the filename in the #include directive I have chosen to change the filename to make it consistent with the other file of the similar name. A search indicates that this is also consistent with the official spelling of the font.